requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
beautifulsoup4
requests
streamlit
pandas
python-dotenv
google-genai
tenacity
httpx
